actually use a 64bits int for invoke test

+++ test/c/invoke_with_symbols.c
@@ -1,6 +1,6 @@
 #include <owi.h>
 
-void f(int n, long int m, float x, double y) {
+void f(int n, long long int m, float x, double y) {
     owi_assume(n == 1);
     owi_assume(y == -1234.);
     owi_assume(n > m);
